+++ app/controllers/avocado/affirmations_controller.rb
@@ -4,20 +4,23 @@ module Avocado
   class AffirmationsController < BaseController
     skip_before_action :authenticate
 
-    before_action :set_user, only: :show
+    before_action :set_user, only: %i[edit update]
     before_action :verify_user, only: :create
 
     def new
     end
 
-    def show
-      sign_in(@user)
-      redirect_to(root_path, notice: "Signed in successfully")
-    end
-
     def create
       send_affirmation_email
       redirect_to new_session_path, notice: "Check your email for sign in instructions"
+    end
+
+    def edit
+    end
+
+    def update
+      sign_in(@user)
+      redirect_to(root_path, notice: "Signed in successfully")
     end
 
     private