+++ spec/system/affirmations_spec.rb
@@ -33,20 +33,32 @@ RSpec.describe "Affirmations" do
 
         travel_to (Avocado::UserTokens::EXPIRES_FAST + 1.day).from_now
 
-        visit affirmation_path(id: user_sid)
+        visit edit_affirmation_path(id: user_sid)
 
         expect(page).to have_content(/is invalid/)
       end
     end
 
     context "with a valid link" do
-      it "signs in the user" do
+      it "signs in the user when they click in time" do
         user = create(:user, verified: false)
         user_sid = user.generate_token_for(:email_affirmation)
 
-        visit affirmation_path(id: user_sid)
+        visit edit_affirmation_path(id: user_sid)
+        click_on "Submit"
 
         expect(page).to have_content(/Signed in/)
+      end
+
+      it "does not sign in the user with a long delay" do
+        user = create(:user, verified: false)
+        user_sid = user.generate_token_for(:email_affirmation)
+
+        visit edit_affirmation_path(id: user_sid)
+        travel_to (Avocado::UserTokens::EXPIRES_FAST + 1.day).from_now
+        click_on "Submit"
+
+        expect(page).to have_content(/is invalid/)
       end
     end
   end